+++ tests/testthat/test-6-hpc.R
@@ -70,11 +70,13 @@ test_with_dir("checksum functionality", {
   config$cache <- decorate_storr(storr::storr_environment())
   testrun(config)
   checksum <- get_checksum(target = "combined", value = 1, config = config)
-  bad <- "askldfklhjsdfkj"
-  expect_false(grepl("NA", checksum))
+  splits <- unlist(strsplit(checksum, split = " "))
+  out <- paste(splits[-3], collapse = " ")
+  expect_false(grepl("NA", out))
   expect_true(
     is_good_checksum(
       target = "combined", value = 1, checksum = checksum, config = config))
+  bad <- "askldfklhjsdfkj"
   expect_false(
     is_good_checksum(
       target = "combined", value = 1, checksum = bad, config = config))